Make rust-clippy happy
